Revert "refactor: change spaceone related words, url, etc. to cloudforet  (#2)"
This reverts commit f9a5265968ba42c03ad4832c990eecd63ea9f899.
