+++ .gitmodules
@@ -1,6 +1,6 @@
 [submodule "src/assets"]
 	path = src/assets
-	url = https://github.com/cloudforet-io/console-assets.git
+	url = https://github.com/spaceone-dev/console-assets.git
 [submodule "src/translations/language-pack"]
 	path = src/translations/language-pack
-	url = https://github.com/cloudforet-io/design-system-translation.git
+	url = https://github.com/spaceone-dev/design-system-translation.git
