+++ .storybook/manager.js
@@ -1,7 +1,7 @@
 window.STORYBOOK_GA_ID='UA-159327743-4' // Google Analytics
 
 import { addons } from '@storybook/addons';
-import SpaceOneTheme from './CloudforetTheme';
+import SpaceOneTheme from './SpaceOneTheme';
 
 addons.setConfig({
     theme: SpaceOneTheme,
